Move deframer tests to their own module to make triaging panics easier.
The triage process filters out code in "_test.rs" files.
