+++ rustls/src/msgs/mod.rs
@@ -8,6 +8,8 @@ pub mod base;
 pub mod ccs;
 pub mod codec;
 pub mod deframer;
+#[cfg(test)]
+mod deframer_test;
 #[allow(non_camel_case_types)]
 pub mod enums;
 pub mod fragmenter;
